requirements.txt
# 戶政門牌爬蟲依賴套件

# HTTP 請求
requests>=2.28.0

# OCR 驗證碼識別
ddddocr>=1.0.6
Pillow>=9.0.0

# MySQL 資料庫
pymysql>=1.0.0
cryptography>=41.0.0

# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0

# 排程
apscheduler>=3.10.0

# 其他
python-dotenv>=1.0.0
